Add TowerTest class with first test
The first test checks that shot is successful when we mock Random object passed
to constructor of Tower class. Mocking is set in private method that takes one
argument: value to be used as thrown by Random.nextDouble(). in [SetUp]
method test Map is set to (5,5).

RandomMock is used as a public member that is verified in tests and set up in
private method